Fix size_t discrepancy

+++ carbon/atmegau2/usb/bulk.c
@@ -4,7 +4,7 @@
 
 
 /* Receive a packet using the appropriate bulk endpoint. */
-int8_t megausb_bulk_receive(void *dst, size_t length) {
+int8_t megausb_bulk_receive(void *dst, uint32_t length) {
 
 	/* If USB is not configured, return with error. */
 	if (!megausb_configuration) {
@@ -37,12 +37,12 @@ int8_t megausb_bulk_receive(void *dst, size_t length) {
 			UENUM = BULK_OUT_ENDPOINT;
 		}
 
-		/* Transfer the buffered data to the destination. */
+		/* Transfer the buffered data to the dst. */
 		uint8_t len = BULK_OUT_SIZE;
 		while (len --) {
 			if (length) {
 				/* If there is still valid data to send, load it from the receive buffer. */
-				*(uint8_t *) dst++ = UEDATX;
+				*(uint8_t *)dst++ = UEDATX;
 				/* Decrement the length. */
 				length --;
 			} else {
@@ -66,7 +66,7 @@ int8_t megausb_bulk_receive(void *dst, size_t length) {
 }
 
 /* Receive a packet using the appropriate bulk endpoint. */
-int8_t megausb_bulk_transmit(void *src, size_t length) {
+int8_t megausb_bulk_transmit(void *src, uint32_t length) {
 
 	/* If USB is not configured, return with error. */
 	if (!megausb_configuration) {
@@ -99,12 +99,12 @@ int8_t megausb_bulk_transmit(void *src, size_t length) {
 			UENUM = BULK_IN_ENDPOINT & ~USB_IN_MASK;
 		}
 
-		/* Transfer the buffered data to the destination. */
+		/* Transfer the buffered data to the dst. */
 		uint8_t len = BULK_IN_SIZE;
 		while (len --) {
 			if (length) {
 				/* If there is still valid data to send, load it into the transmit buffer. */
-				UEDATX = *(uint8_t *) src++;
+				UEDATX = *(uint8_t *)src++;
 				/* Decrement the length. */
 				length --;
 			} else {
