+++ carbon/include/flipper/libc.h
@@ -3,7 +3,7 @@
 
 #include <flipper/types.h>
 
-void *libc_malloc(size_t length);
+void *libc_malloc(uint32_t length);
 void libc_free(void *ptr);
 void *libc_memcpy(void *dst, void *src, size_t n);
